collapse game and gamestate
